requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
groq
llama-cpp-python
huggingface-hub
ollama
